+++ qcio/models/single_point.py
@@ -1,13 +1,15 @@
 from enum import Enum
-from typing import TYPE_CHECKING, Optional, Type, Union
+from typing import TYPE_CHECKING, Any, Dict, Optional, Type, Union
 
 import numpy as np
 from pydantic import BaseModel, validator
 
-from ..helper_types import ArrayLike2D
+from qcio.helper_types import ArrayLike2D
+
 from .base_io import (
     ComputedPropertiesBase,
     FailedOutputBase,
+    Provenance,
     StructuredInputBase,
     StructuredProgramArgsBase,
     SuccessfulOutputBase,
@@ -15,7 +17,7 @@ from .base_io import (
 from .base_model import QCIOModelBase
 from .molecule import Molecule
 
-if TYPE_CHECKING:
+if TYPE_CHECKING:  # pragma: no cover
     from pydantic.typing import ReprArgs
 
 __all__ = [
@@ -39,7 +41,7 @@ class SPCalcType(str, Enum):
 
     def __repr__(self) -> str:
         """Custom repr for CalcType"""
-        return f"'{self.name}'"
+        return f"'{self.name}'"  # pragma: no cover
 
 
 class Model(BaseModel):
@@ -47,7 +49,8 @@ class Model(BaseModel):
 
     Attributes:
         method: The name of the method to be used in the calculation. Named according to
-            the convention of the program being called.
+            the convention of the program being called. If an MM calculation the name of
+            the force field.
         basis: The name of the basis set to be used in the calculation. Named according
             to the convention of the program being called.
     """
@@ -75,7 +78,7 @@ class SinglePointProgramArgs(StructuredProgramArgsBase):
 
     def __repr_args__(self) -> "ReprArgs":
         """Custom repr for SinglePointProgramArgs"""
-        return [
+        return [  # pragma: no cover
             ("calc_type", self.calc_type),
             ("model", self.model),
         ]
@@ -140,14 +143,16 @@ class SinglePointComputedProperties(ComputedPropertiesBase):
     @validator("gradient")
     def validate_gradient_shape(cls, v: ArrayLike2D):
         """Validate gradient is n x 3"""
-        return np.asarray(v).reshape(-1, 3)
+        if v is not None:
+            return np.asarray(v).reshape(-1, 3)
 
     @validator("hessian")
     def validate_hessian_shape(cls, v: ArrayLike2D):
         """Validate hessian is square"""
-        v = np.asarray(v)
-        n = int(np.sqrt(v.size))
-        return v.reshape((n, n))
+        if v is not None:
+            v = np.asarray(v)
+            n = int(np.sqrt(v.size))
+            return v.reshape((n, n))
 
 
 class SinglePointOutputMixin(BaseModel):
@@ -158,6 +163,133 @@ class SinglePointOutputMixin(BaseModel):
     """
 
     input_data: "SinglePointInput"
+
+    def __repr_args__(self) -> "ReprArgs":
+        """A helper for __repr__ that returns a list of tuples of the form
+        (name, value).
+        """
+        return [  # pragma: no cover
+            ("calc_type", self.input_data.program_args.calc_type),
+            ("model", self.input_data.program_args.model),
+            ("formula", self.input_data.molecule.formula),
+        ]
+
+
+class SinglePointInput(StructuredInputBase):
+    """Input for a single point calculation.
+
+    Attributes:
+        program_args: Input arguments for the program.
+        molecule: The molecule to be used in the calculation.
+        extras: Additional information to bundle with the object. Use for schema
+            development and scratch space.
+    """
+
+    program_args: SinglePointProgramArgs
+    molecule: Molecule
+
+    def __repr_args__(self) -> "ReprArgs":
+        return [  # pragma: no cover
+            ("calc_type", self.program_args.calc_type),
+            ("model", self.program_args.model),
+            ("keywords", self.program_args.keywords),
+            ("formula", self.molecule.formula),
+        ]
+
+    def get_failed_output_class(self) -> Type["SinglePointFailedOutput"]:
+        """Return the FailedOutput class for this input class."""
+        return SinglePointFailedOutput
+
+    def get_successful_output_class(self) -> Type[SuccessfulOutputBase]:
+        return SinglePointSuccessfulOutput
+
+    def to_qcel(self) -> Dict[str, Any]:
+        """Return the QCElemental v1 input schema representation of the input
+        (AtomicInput dict).
+
+        Returns:
+            The QCElemental v1 dict representation of an AtomicInput object.
+        """
+        return {
+            "molecule": {
+                "symbols": self.molecule.symbols,
+                "geometry": self.molecule.geometry,
+                "molecular_charge": self.molecule.charge,
+                "molecular_multiplicity": self.molecule.multiplicity,
+                "identifiers": {
+                    key: value
+                    for key, value in self.molecule.identifiers.dict().items()
+                    if key not in ["name_IUPAC", "name_common"]  # not on qcel model
+                },
+            },
+            "driver": self.program_args.calc_type,
+            "model": self.program_args.model.dict(),
+            "keywords": self.program_args.keywords,
+            "extras": self.extras,
+        }
+
+    def to_output_from_qcel(
+        self,
+        qcel_output: Dict[str, Any],
+    ) -> Union["SinglePointSuccessfulOutput", "SinglePointFailedOutput"]:
+        """Create a SinglePointSuccessfulOutput or SinglePointFailedOutput from the
+        QCElemental v1 output schema representation of the output (AtomicResult dict).
+
+        Args:
+            provenance: The qcio Provenance object for the output.
+            qcel_output: The QCElemental v1 output schema representation of the output.
+                May be a dict representing an AtomicResult or FailedOperation.
+        """
+        prov_extras = {"NOTE": "Computed by QCEngine"}
+
+        if qcel_output["success"] is False:
+            return SinglePointFailedOutput(
+                input_data=self,
+                traceback=qcel_output["error"]["error_message"],
+                provenance=Provenance(
+                    program=qcel_output["input_data"]["provenance"]["creator"],
+                    program_version=qcel_output["input_data"]["provenance"].get(
+                        "version"
+                    ),
+                    wall_time=qcel_output["input_data"]["provenance"].get("wall_time"),
+                    extras=prov_extras,
+                ),
+                extras=qcel_output["extras"] or {},  # Because .extras may be None
+            )
+        else:
+            # Collect values from keys that exist in qcio
+            qcio_to_qcel = {
+                "calcinfo_natoms": "calcinfo_natom",
+                "energy": "return_energy",
+                "gradient": "return_gradient",
+                "hessian": "return_hessian",
+            }
+            computed = {}
+            for key in SinglePointComputedProperties.__annotations__:
+                if key in qcio_to_qcel:
+                    qcel_key = qcio_to_qcel[key]
+                else:
+                    qcel_key = key
+                computed[key] = qcel_output["properties"].get(qcel_key)
+
+            if qcel_output["wavefunction"]:
+                computed["wfn"] = {
+                    key: value
+                    for key, value in qcel_output["wavefunction"].items()
+                    if key in Wavefunction.__annotations__
+                }
+
+            return SinglePointSuccessfulOutput(
+                input_data=self,
+                stdout=qcel_output["stdout"],
+                computed=SinglePointComputedProperties(**computed),
+                provenance=Provenance(
+                    program=qcel_output["provenance"]["creator"],
+                    program_version=qcel_output["provenance"].get("version"),
+                    wall_time=qcel_output["provenance"].get("wall_time"),
+                    extras=prov_extras,
+                ),
+            )
 
 
 class SinglePointFailedOutput(SinglePointOutputMixin, FailedOutputBase):
@@ -176,34 +308,7 @@ class SinglePointFailedOutput(SinglePointOutputMixin, FailedOutputBase):
             development and scratch space.
     """
 
-    input_data: "SinglePointInput"
-
-
-class SinglePointInput(StructuredInputBase):
-    """Input for a single point calculation.
-
-    Attributes:
-        program_args: Input arguments for the program.
-        molecule: The molecule to be used in the calculation.
-        extras: Additional information to bundle with the object. Use for schema
-            development and scratch space.
-    """
-
-    program_args: SinglePointProgramArgs
-    molecule: Molecule
-
-    def __repr_args__(self) -> "ReprArgs":
-        return [
-            ("molecule", self.molecule),
-            ("program_args", self.program_args),
-        ]
-
-    def get_failed_output_class(self) -> Type[SinglePointFailedOutput]:
-        """Return the FailedOutput class for this input class."""
-        return SinglePointFailedOutput
-
-    def get_successful_output_class(self) -> Type[SuccessfulOutputBase]:
-        return SinglePointSuccessfulOutput
+    input_data: SinglePointInput
 
 
 class SinglePointSuccessfulOutput(SinglePointOutputMixin, SuccessfulOutputBase):
@@ -223,14 +328,11 @@ class SinglePointSuccessfulOutput(SinglePointOutputMixin, SuccessfulOutputBase):
     computed: SinglePointComputedProperties
 
     def __repr_args__(self) -> "ReprArgs":
-        """A helper for __repr__ that returns a list of tuples of the form
-        (name, value).
-        """
-        return [
-            ("success", self.success),
-            ("input_data", self.input_data),
-            ("computed", "(...)"),
-            ("provenance", self.provenance),
+        return super().__repr_args__() or [] + [  # pragma: no cover
+            (
+                "energy",
+                round(self.computed.energy, 6) if self.computed.energy else None,
+            ),
         ]
 
     @property